Split direct connectivity diagnostics from mesh
